+++ src/main/java/dip/clever/controller/MypageController.java
@@ -1,0 +1,32 @@
+package dip.clever.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class MypageController {
+
+	// mypage 반환
+	@RequestMapping("/mypage")
+	public String mypage() {
+		return "mypage/mypage";
+	}
+
+	// mypage - 개인정보 수정
+	@RequestMapping("/mypage-setting")
+	public String mypageSetting() {
+		return "mypage/settings/mypage-setting";
+	}
+
+	// mypage - 프로필/계정 정보 수정
+	@PostMapping("/settings-{category}")
+	public String settingsProfile(@PathVariable String category) {
+		if (category.equals("profile")) {
+			return "mypage/settings/settings-profile";
+		} else {
+			return "mypage/settings/settings-account";
+		}
+	}
+}
